WIP: 임시저장

+++ Source/Public/CLogManager.h
@@ -15,7 +15,7 @@ private:
 	std::vector<std::unique_ptr<ILogable>> logs;
 
 public:
-	inline int GetLogCount() const { return logs.size(); }
+	inline int GetLogCount() const { return (int)logs.size(); }
 	inline const std::vector<std::unique_ptr<ILogable>>& GetLogs() const { return logs; }
 	int AddLog(std::unique_ptr<ILogable> log);
